% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_SF.R
\name{Bensh_RSF}
\alias{Bensh_RSF}
\title{Bensh_RSF Function}
\usage{
Bensh_RSF(rsf36, rsf12, layout, power_curve)
}
\arguments{
\item{rsf36:}{first rsf object already loaded into memory}

\item{rsf12:}{second rsf object already loaded into memory}

\item{layout:}{data frame of two columns containing the turbine positions (X,Y).}

\item{power_curve:}{power curve used to compute production. the format of power_curve is a data frame with two columns (Wind speed [m/s], power [kW]). can be generate by Gen_power_curve function.}
}
\description{
Benshark wind turbine production of two different rsf (wind mapping). It is mainly use to benchmark 36 versus 12 sectors rsf.
}
\examples{
Compute AEP for V117 on the position (3756122,1625482):
layout <- read.csv("layout_file.csv")
pc <- Gen_power_curve("V117-3.45")
bench_table <- Bensh_RSF(rsf_H1_36,rsf_H1_12, layout, pc)
}
\keyword{benshmark}
\keyword{production,}
\keyword{rsf,}

