% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_SF.R
\name{Gen_power_curve}
\alias{Gen_power_curve}
\title{Gen_power_curve Function}
\usage{
Gen_power_curve(turbine = "V117-3.45")
}
\arguments{
\item{turbine:}{choose the turbine type. (for the moment only "V117-3.45" is availbale)}
}
\description{
Generate a power curve. the format of the generated power_curve is a data frame with two columns (Wind speed [m/s], power [kW])
}
\examples{
generate V117 power curve and store it in "pc" object:
pc <- Gen_power_curve("V117-3.45")
}
\keyword{curve}
\keyword{power}
\keyword{production,}
\keyword{rsf,}

