% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignore_outputs.R
\name{git_ignore_outputs}
\alias{git_ignore_outputs}
\title{Git ignore generated outputs}
\usage{
git_ignore_outputs(figs_only = TRUE)
}
\arguments{
\item{figs_only}{logical; if \code{TRUE} (default), only figures are git ignored}
}
\description{
Add generated reports, figures, and HTML libraries to .gitignore
}
\examples{
\dontrun{
git_ignore_outputs(figs_only = FALSE)
}
}
