% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_SF.R
\name{Prod_RSF}
\alias{Prod_RSF}
\title{Prod_RSF Function}
\usage{
Prod_RSF(rsf, Xp, Yp, power_curve)
}
\arguments{
\item{rsf:}{rsf object already loaded into memory}

\item{Xp:}{turbine position (x)}

\item{Yp:}{turbine position (y)}

\item{power_curve:}{power curve used to compute production. the format of power_curve is a data frame with two columns (Wind speed [m/s], power [kW]). can be generate by Gen_power_curve() function.}
}
\description{
Compute the production [MWh] for one turbine type (power_curve) at the unique position (Xp, Yp) based on the rsf (rsf).
}
\examples{
Compute AEP for V117 at the position (3756122,1625482) and store it into AEP_T1:
pc <- Gen_power_curve("V117-3.45")
AEP_T1 <- Prod_RSF(rsf87,3756122,1625482,pc)
}
\keyword{production}
\keyword{rsf,}

