% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_SF.R
\name{RSF36_2_12}
\alias{RSF36_2_12}
\title{RSF36_2_12 Function}
\usage{
RSF36_2_12(rsf_file, layout_csv, output_name = rsf_file1)
}
\arguments{
\item{Input_file:}{"*.rsf" file ()}
}
\description{
Based on 2 inputs files (36 Sect rsf, and layout (*.csv)), this "Master" function generate a "Cropped" rsf (700m around the turbine positions)
and also a "Cropped + 12 sectors"" rsf. this "Master" function taking into account the input files as direct input (and then call the other sub functions)
}
\examples{
load into memory an rsf file into the object rsf_166:
rsf_166 <- Read_RSF("RSF-windresource-CFD_Aldermyrberget 166m.rsf")
}
\keyword{read}
\keyword{rsf,}

