% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_SF.R
\name{RSF_scale}
\alias{RSF_scale}
\title{RSF_scale}
\usage{
RSF_scale(rsf_file, scale_factor, output_name = rsf_file)
}
\arguments{
\item{rsf_file:}{Input *.rsf file name (name should be surrounded by "").}

\item{scale_factor:}{the A value of the rsf will be multiply by this value.}

\item{output_name:}{Output *.rsf file name " (name should be surrounded by "").}
}
\description{
This function allows to add scale a rsf with a factor x.
The new rsf file will be created in the same folder, with the scale factor tagged in the file name.
Note that the scale is purely apply to the A parameter of the weibull distribution (k stay unchanged)
}
\examples{
Scale the rsf file "RSF-CFD_Douglas_West.rsf" by 1.11 (add 11 percents to the wind speed)  :
RSF_scale(rsf_file_file = "RSF-CFD_Douglas_West.rsf",scale = 1.11)
}
\keyword{rsf,}
\keyword{scale}
