% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_SF.R
\name{S36_to_S12_RSF}
\alias{S36_to_S12_RSF}
\title{S36_to_S12_RSF Function}
\usage{
S36_to_S12_RSF(RSF36)
}
\arguments{
\item{RSF36:}{36 sectors rsf (already loaded in memory by the use fo Read_RSF() ).}
}
\description{
Tranform rsf from 36 sectors to 12. it combines the neighboring sectors 3 by 3 by doing a ponderate average.
}
\examples{
convert rsf87_36S from 36 to 12 sector and store the result in rsf87_12S:
rsf87_12S <- S36_to_S12_RSF(rsf87_36S)
}
\keyword{36,}
\keyword{convertion}
\keyword{rsf,}

