% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_SF.R
\name{Shear2png}
\alias{Shear2png}
\title{Shear2png Function}
\usage{
Shear2png(RSF_shear)
}
\arguments{
\item{RSF_shear:}{"shear rsf" already loaded into memory (could be generated by the function Shear_RSF())}
}
\description{
Export the "RSF_shear" rsf to a serie of georefenced images (*.png and *.pgw). For more information about the *.pgw (word file): https://en.wikipedia.org/wiki/World_file). those images can be directly used in windpro as background maps.
}
\examples{
generate a shear rsf from rsf_87 and rsf137 and export it into png pictures.
shear_ <- Shear_RSF(rsf87,rsf137)
Shear2png(shear_)
}
\keyword{png}
\keyword{rsf,}
\keyword{shear,}

