% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_SF.R
\name{Shear2pngV2}
\alias{Shear2pngV2}
\title{Shear2pngV2 Function}
\usage{
Shear2pngV2(RSF_shear)
}
\arguments{
\item{RSF_shear:}{"Shear rsf" already loaded into memory}
}
\description{
Save (export) the "RSF_shear" rsf to a series (average and for all sectors) of georefenced images (*.png and *.pgw). for more information about the *.pgw (word file): https://en.wikipedia.org/wiki/World_file). Those images can be directly used in windpro as background maps.
}
\examples{
generate a shear rsf from rsf_87 and rsf137 and export it into png pictures.
shear_ <- Shear_RSF(rsf87,rsf137)
Shear2pngV2(shear_)
}
\keyword{png}
\keyword{rsf,}
\keyword{shear,}
