% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_SF.R
\name{ShearExtract}
\alias{ShearExtract}
\title{ShearExtract Function}
\usage{
ShearExtract(shear_csv, layout_csv, output_name = "shear_VSC_input.shr")
}
\arguments{
\item{shear_csv:}{shear file in CSV format (output of ShearTable() function)}

\item{layout_csv:}{layout file in csv format (X,Y)}

\item{output_name:}{name of the output file,it has to contain the extension "*.shr"}
}
\description{
Extract shear at turbine locations, and create a *.shr file for VSC (use as CFD input)
}
\examples{
Exract directional shear for each turbine location present in the file "layout_file.csv".
shear_ <- Shear_RSF(rsf87,rsf137)
ShearExtract("Sheartable.csv","layout_file.csv","shear_VSC_input.shr")
}
\keyword{extraction}
\keyword{rsf,}
\keyword{shear,}

