% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_rsf.R
\name{use_rsf}
\alias{use_rsf}
\title{Use RSF project template}
\usage{
use_rsf(path, ...)
}
\arguments{
\item{path}{project path}

\item{...}{project configurations supported:
\itemize{
\item initialize git repo via \code{\link[gert:git_init]{gert::git_init()}}
\item initialize renv via \code{\link[renv:init]{renv::init()}}
}}
}
\description{
Create a new Quarto book R project using the RSF template. Intended for use
in RStudio menu, not interactively.
}
\details{
This function is called when the user selects File > New Project > New
Directory > Report of Statistical Findings using Quarto. The directory name
can be specified, and the user can choose to initialize the project as
a git repository and/or use with \code{renv}.
}
