% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_SF.R
\name{Vizu_RSF}
\alias{Vizu_RSF}
\title{Vizu_RSF Function}
\usage{
Vizu_RSF(rsf)
}
\arguments{
\item{rsf:}{first rsf object already loaded into memory (wind mapping)}
}
\description{
Interactive plot of a rsf object (direction, Height, parameters...)
}
\examples{
Vizu_RSF(rsf_87)
}
\keyword{plot,}
\keyword{rsf,}
\keyword{vizualisation}

