% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_SF.R
\name{Write_RSF}
\alias{Write_RSF}
\title{Write_RSF Function}
\usage{
Write_RSF(RSF, output_file_name)
}
\arguments{
\item{RSF:}{rsf object already loaded into memory}

\item{output_file_name:}{name for the output file. it has to contain the extension .rsf.}
}
\description{
Write a rsf object from memory to a file (*.rsf). This file can be used with Windpro.
}
\examples{
write from memory a rsf (data table) into the file "rsf_80m_Aldermyberget.rsf":
Write_RSF(rsf80, "rsf_80m_Aldermyberget.rsf")
}
\keyword{rsf,}
\keyword{write}
