% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsf-package.R
\docType{package}
\name{rsf-package}
\alias{rsf}
\alias{rsf-package}
\title{rsf: Report of Statistical Findings in 'bookdown'}
\description{
A report of statistical findings (RSF) project template is generated using a 'bookdown' format. 'YAML' fields can be further customized. Additional helper functions provide extra features to the RSF.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dchiu911/rsf/}
  \item \url{https://dchiu911.github.io/rsf/}
  \item Report bugs at \url{https://github.com/dchiu911/rsf/issues}
}

}
\author{
\strong{Maintainer}: Derek Chiu \email{dchiu@bccrc.ca}

}
\keyword{internal}
