% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_references.R
\name{use_references}
\alias{use_references}
\title{Use references}
\usage{
use_references(path = ".", number = 99)
}
\arguments{
\item{path}{path of the directory containing the input files \code{index.Rmd},
\code{_output.yml}, \code{preamble.tex}.}

\item{number}{number prepended to references Rmd source file.}
}
\description{
Add a References section at the end of the RSF.
}
\examples{
\donttest{
wd <- tempdir()
savedir <- setwd(wd)
file.copy(list.files(system.file("extdata", package = "rsf"),
                     full.names = TRUE), ".")
use_references(number = 4)
setwd(savedir)
}
}
