% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_rsf.R
\name{use_rsf}
\alias{use_rsf}
\title{Use RSF project template}
\usage{
use_rsf(path, ...)
}
\arguments{
\item{path}{project path}

\item{...}{project configurations supported:
\itemize{
\item \code{output_dir} in \verb{_bookdown.yml} \url{https://bookdown.org/yihui/bookdown/configuration.html}
\item initialize git repo via \code{\link[gert:git_repo]{gert::git_init()}}
\item initialize renv via \code{\link[renv:init]{renv::init()}}
}}
}
\description{
Opens a new R project using the RSF template. Intended for use in RStudio,
not interactively.
}
\details{
This function is called when the user selects File > New Project > New
Directory > Report of Statistical Findings using bookdown. The directory name
and output directory can be specified.
}
